[Rename] sub-sidebar 하위 디렉토리 구조 변경
