+++ components/layouts/sidebar/footer/tier-info-modal.tsx
@@ -13,18 +13,6 @@ interface TierInfoModalProps {
   handleClose: () => void;
 }
 
-const style = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: '30%',
-  height: '65%',
-  bgcolor: 'white',
-  boxShadow: 24,
-  p: 4,
-};
-
 function Badge({ tier, range, color }: BadgeProps) {
   return (
     <div className="flex items-center gap-3">
