+++ components/layouts/sidebar/sub-sidebar/collected-card/detail.tsx
@@ -6,9 +6,9 @@ import {
   SubsidebarCloseIconStyle,
 } from 'utils/styles';
 import { CollectedCafeDetailProps } from 'types/types';
-import OpenTimeGrid from './open-time-grid';
-import LocationGrid from './location-grid';
-import PhoneGrid from './phone-grid';
+import OpenTimeGrid from '../common/open-time-grid';
+import LocationGrid from '../common/location-grid';
+import PhoneGrid from '../common/phone-grid';
 import Image from 'next/image';
 import ConceptGrid from './concept-grid';
 import EatenGrid from './eaten-grid';
@@ -28,7 +28,7 @@ export default function CollectedCafeDetail({
     (state: any) => state.collectedCafeDetail[0]
   );
 
-  // 서브 사이드바에서 넘겨준 set 함수들로 메모 인풋의 값들 설정하기
+  // 서브 사이드바에서 넘겨준 set 함수들로 메모 인풋의 value로 설정하기
 
   return (
     <div className={`flex flex-col p-2 gap-4`}>
