+++ components/layouts/sidebar/sub-sidebar/collected-card/eaten-grid.tsx
