+++ components/layouts/sidebar/sub-sidebar/collected-card/rating-grid.tsx
