+++ components/layouts/sidebar/sub-sidebar/normal-card/collected-badge.tsx
