+++ components/layouts/sidebar/sub-sidebar/normal-card/detail.tsx
@@ -15,12 +15,12 @@ import {
 } from 'utils/styles';
 import { toast } from 'react-toastify';
 import { NormalCafeDetailProps } from 'types/types';
-import CollectedBadge from 'components/layouts/sidebar/sub-sidebar/collected-badge';
+import CollectedBadge from 'components/layouts/sidebar/sub-sidebar/normal-card/collected-badge';
 import ReviewAndRatingGrid from './review-and-rating-grid';
 import Image from 'next/image';
-import OpenTimeGrid from './open-time-grid';
-import LocationGrid from './location-grid';
-import PhoneGrid from './phone-grid';
+import OpenTimeGrid from '../common/open-time-grid';
+import LocationGrid from '../common/location-grid';
+import PhoneGrid from '../common/phone-grid';
 import MenuGrid from './menu-grid';
 import BookmarkIcon from '@mui/icons-material/Bookmark';
